requirements.txt
### requirements.txt


transformers
Pillow
numpy
